Rewrite TestProcessHooks from scratch
